close

+++ lib/components/rdv_tab_med_portail.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_svg/svg.dart';
+import 'package:medstory/components/rdv_reprog.dart';
 import 'package:medstory/models/medecin.dart';
 import 'package:medstory/models/my_data.dart';
 import 'package:medstory/models/rendez_vous.dart';
@@ -19,9 +20,22 @@ class RdvTabMedPortail extends StatefulWidget {
 class _RdvTabMedPortailState extends State<RdvTabMedPortail> {
   final rendezVousService = RendezVousService();
   Medecin? medecin;
+
+  @override
+  void initState() {
+    super.initState();
+    WidgetsBinding.instance.addPostFrameCallback((_) {
+      context
+          .read<MyData>()
+          .fetchRendezVousmedecin(context.watch<MyData>().currentMedecin!.id!);
+    });
+  }
+
   @override
   Widget build(BuildContext context) {
-    medecin = context.watch<MyData>().currentMedecin!;
+    medecin = context
+        .watch<MyData>()
+        .currentMedecin!; //TODO: s'assurer d'avoir le bon
     return Column(
       children: [
         DataTable(
@@ -90,21 +104,30 @@ class _RdvTabMedPortailState extends State<RdvTabMedPortail> {
                           icon: const Icon(Icons.cancel, color: Colors.orange),
                           onPressed: () async {
                             // Action pour annuler le rendez-vous
-                            context.showLoader();
-
-                            await rendezVousService.changeRendezVousStatut(
-                                rendezVous.id,
-                                {"id": 2, "libelle": "annulé"}).then((value) {
+                            customRdvModal(
+                              context,
+                              rendezVous,
+                            ).whenComplete(() {
                               context
                                   .read<MyData>()
                                   .fetchRendezVousmedecin(medecin!.id!);
-                              context.hideLoader();
-                              context.showSuccess(
-                                  "Le rendez-vous a été annulé avec succès.");
-                            }).catchError((onError) {
-                              context.hideLoader();
-                              context.showError("Oups !");
                             });
+
+                            // context.showLoader();
+
+                            // await rendezVousService.changeRendezVousStatut(
+                            //     rendezVous.id,
+                            //     {"id": 2, "libelle": "annulé"}).then((value) {
+                            //   context
+                            //       .read<MyData>()
+                            //       .fetchRendezVousmedecin(medecin!.id!);
+                            //   context.hideLoader();
+                            //   context.showSuccess(
+                            //       "Le rendez-vous a été annulé avec succès.");
+                            // }).catchError((onError) {
+                            //   context.hideLoader();
+                            //   context.showError("Oups !");
+                            // });
                           },
                         ),
                   IconButton(
@@ -115,36 +138,66 @@ class _RdvTabMedPortailState extends State<RdvTabMedPortail> {
                     ),
                     onPressed: () async {
                       // Action pour supprimer le médecin
-                      context.showLoader();
-                      final rendezVousService = RendezVousService();
-                      await rendezVousService
-                          .deleteRendezVous(rendezVous.id)
-                          .then((value) {
-                        context
-                            .read<MyData>()
-                            .fetchRendezVous(); //TODO: Remplacer la logique (fetch only for current user)
-                        context.hideLoader();
-                        context.showSuccess(
-                            "Le rendez-vous a été supprimé avec succès.");
-                      }).catchError((onError) {
-                        context.hideLoader();
-                        context.showError(onError.toString());
-                      });
+                      context.showConfirmation(
+                        title: "Attention !!",
+                        message:
+                            "Cette action supprime le rendez-vous.\nÊtes-vous sûr(e) de continuer ?",
+                        onConfirm: () async {
+                          context.showLoader();
+                          final rendezVousService = RendezVousService();
+                          await rendezVousService
+                              .deleteRendezVous(rendezVous.id)
+                              .then((value) {
+                            context.read<MyData>().fetchRendezVousmedecin(medecin!
+                                .id!); //TODO: Remplacer la logique (fetch only for current user)
+                            context.hideLoader();
+                            context.showSuccess(
+                                "Le rendez-vous a été supprimé avec succès.");
+                          }).catchError((onError) {
+                            context.hideLoader();
+                            context.showError(onError.toString());
+                          });
+                        },
+                      );
                     },
                   ),
                 ],
               )),
-              DataCell(Text(
-                  '${rendezVous.patient.prenom} ${rendezVous.patient.nom}')),
-              DataCell(Text(
-                  '${rendezVous.date.day}/${rendezVous.date.month}/${rendezVous.date.year}')),
-              DataCell(Text(rendezVous.heure)),
-              DataCell(Text(rendezVous.motif)),
-              DataCell(Text(rendezVous.statut.libelle)),
+              DataCell(SizedBox(
+                width: 200,
+                child: Text(
+                    '${rendezVous.patient.prenom} ${rendezVous.patient.nom}'),
+              )),
+              DataCell(SizedBox(
+                width: 100,
+                child: Text(
+                    '${rendezVous.date.day}/${rendezVous.date.month}/${rendezVous.date.year}'),
+              )),
+              DataCell(SizedBox(width: 100, child: Text(rendezVous.heure))),
+              DataCell(SizedBox(width: 100, child: Text(rendezVous.motif))),
+              DataCell(
+                  SizedBox(width: 100, child: Text(rendezVous.statut.libelle))),
             ]);
           }).toList(),
         ),
       ],
     );
+  }
+
+  Future<dynamic> customRdvModal(BuildContext contexte, RendezVous rdv) {
+    return showDialog(
+        context: contexte,
+        builder: (contexte) {
+          return Material(
+            type: MaterialType.transparency,
+            child: FractionallySizedBox(
+              heightFactor: 0.85,
+              widthFactor: 0.8,
+              child: RdvReprog(
+                rdv: rdv,
+              ),
+            ),
+          );
+        });
   }
 }
